@@ -2,8 +2,8 @@ package com.example.andrasnemeth.nnotification;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.NotificationCompat;
-import android.support.v4.app.RemoteInput;
+import androidx.core.app.NotificationCompat;
+import androidx.core.app.RemoteInput;
 import android.view.View;
 import android.widget.Toast;
 
